adding Azure Function timer trigger
