+++ source/test_function.py
@@ -1,0 +1,2 @@
+def test_print():
+    print("Testing Azure Function")
